pyproject.toml
[tool.poetry]
name = "meteor"
version = "0.1.0"
description = "Weather Server"
authors = ["macca717 <hayden.mcmenamin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.75.1"
pyhumps = "3.5.3"
aiohttp = "3.8.1"
uvicorn = "0.17.6"
xmltodict = "0.12.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^7.1.1"
requests = "^2.27.1"
anyio = "^3.5.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
